+++ components/selfupdater/api/src/main/kotlin/com/flipperdevices/selfupdater/models/SelfUpdateResult.kt
@@ -1,0 +1,6 @@
+package com.flipperdevices.selfupdater.models
+
+enum class SelfUpdateResult {
+    SUCCESS,
+    NO_UPDATES
+}