+++ components/selfupdater/api/src/main/kotlin/com/flipperdevices/selfupdater/models/SelfUpdateSource.kt
@@ -1,0 +1,8 @@
+package com.flipperdevices.selfupdater.models
+
+enum class SelfUpdateSource {
+    GOOGLE_PLAY,
+    GITHUB,
+    DEBUG,
+    UNKNOWN
+}