+++ components/selfupdater/googleplay/src/main/kotlin/com/flipperdevices/selfupdater/googleplay/api/SelfUpdaterGooglePlay.kt
@@ -1,7 +1,7 @@
 package com.flipperdevices.selfupdater.googleplay.api
 
-import android.app.Activity
 import android.content.Context
+import com.flipperdevices.core.activityholder.CurrentActivityHolder
 import com.flipperdevices.core.di.AppGraph
 import com.flipperdevices.core.log.LogTagProvider
 import com.flipperdevices.core.log.info
@@ -9,6 +9,7 @@ import com.flipperdevices.inappnotification.api.InAppNotificationStorage
 import com.flipperdevices.inappnotification.api.model.InAppNotification
 import com.flipperdevices.selfupdater.api.BuildConfig
 import com.flipperdevices.selfupdater.api.SelfUpdaterApi
+import com.flipperdevices.selfupdater.models.SelfUpdateResult
 import com.google.android.play.core.appupdate.AppUpdateInfo
 import com.google.android.play.core.appupdate.AppUpdateManagerFactory
 import com.google.android.play.core.install.InstallState
@@ -17,6 +18,7 @@ import com.google.android.play.core.install.model.AppUpdateType
 import com.google.android.play.core.install.model.InstallStatus
 import com.google.android.play.core.install.model.UpdateAvailability
 import com.squareup.anvil.annotations.ContributesBinding
+import kotlinx.coroutines.tasks.await
 import javax.inject.Inject
 
 private const val UPDATE_CODE = 228
@@ -44,24 +46,32 @@ class SelfUpdaterGooglePlay @Inject constructor(
         )
     }
 
-    override fun startCheckUpdateAsync(activity: Activity) {
+    override suspend fun startCheckUpdate(onEndCheck: suspend (SelfUpdateResult) -> Unit) {
+        val activity = CurrentActivityHolder.getCurrentActivity() ?: return
+
         info { "Process checkout new update" }
         appUpdateManager.registerListener(updateListener)
-        appUpdateInfoTask.addOnSuccessListener { appUpdateInfo ->
-            info { "Current state update id ${appUpdateInfo.updateAvailability()}" }
-            if (isUpdateAvailable(appUpdateInfo)) {
-                info { "New update available, suggest to update" }
-                appUpdateManager.startUpdateFlowForResult(
-                    appUpdateInfo,
-                    AppUpdateType.FLEXIBLE,
-                    activity,
-                    UPDATE_CODE
-                )
-            }
+
+        val appUpdateInfo = appUpdateInfoTask.await()
+        info { "Current state update id ${appUpdateInfo.updateAvailability()}" }
+
+        if (isUpdateAvailable(appUpdateInfo)) {
+            info { "New update available, suggest to update" }
+            appUpdateManager.startUpdateFlowForResult(
+                appUpdateInfo,
+                AppUpdateType.FLEXIBLE,
+                activity,
+                UPDATE_CODE
+            )
+            onEndCheck(SelfUpdateResult.SUCCESS)
+        } else {
+            info { "No update available" }
+            onEndCheck(SelfUpdateResult.NO_UPDATES)
         }
     }
 
     override fun getInstallSourceName() = "Google Play/" + BuildConfig.BUILD_TYPE
+    override fun isSelfUpdateChecked(): Boolean = false
 
     private fun isUpdateAvailable(appUpdateInfo: AppUpdateInfo): Boolean {
         return appUpdateInfo.updateAvailability() == UpdateAvailability.UPDATE_AVAILABLE &&