+++ components/selfupdater/unknown/src/main/kotlin/com/flipperdevices/selfupdater/unknown/api/SelfUpdaterUnknown.kt
@@ -1,14 +1,15 @@
 package com.flipperdevices.selfupdater.unknown.api
 
-import android.app.Activity
 import com.flipperdevices.core.di.AppGraph
 import com.flipperdevices.selfupdater.api.SelfUpdaterApi
+import com.flipperdevices.selfupdater.models.SelfUpdateResult
 import com.squareup.anvil.annotations.ContributesBinding
 import javax.inject.Inject
 
 @ContributesBinding(AppGraph::class, SelfUpdaterApi::class)
 class SelfUpdaterUnknown @Inject constructor() : SelfUpdaterApi {
-    override fun startCheckUpdateAsync(activity: Activity) = Unit
+    override suspend fun startCheckUpdate(onEndCheck: suspend (SelfUpdateResult) -> Unit) = Unit
 
     override fun getInstallSourceName() = "Unknown"
+    override fun isSelfUpdateChecked(): Boolean = false
 }