+++ components/singleactivity/impl/src/main/java/com/flipperdevices/singleactivity/impl/SingleActivity.kt
@@ -27,6 +27,7 @@ import com.flipperdevices.singleactivity.impl.composable.ComposableSingleActivit
 import com.flipperdevices.singleactivity.impl.di.SingleActivityComponent
 import com.flipperdevices.singleactivity.impl.utils.AppOpenMetricReported
 import kotlinx.collections.immutable.toPersistentSet
+import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.launch
 import javax.inject.Inject
 
@@ -67,7 +68,9 @@ class SingleActivity :
 
         WindowCompat.setDecorFitsSystemWindows(window, false)
 
-        selfUpdaterApi.startCheckUpdateAsync(activity = this@SingleActivity)
+        lifecycleScope.launch(Dispatchers.Default) {
+            selfUpdaterApi.startCheckUpdate(onEndCheck = {})
+        }
 
         val featureEntries = featureEntriesMutable.toPersistentSet()
         val composableEntries = composableEntriesMutable.toPersistentSet()